+++ admin/assets/functions/queries-admin.php
@@ -27,5 +27,7 @@ $categories_portfolio = $bdd->query('SELECT * FROM categories_portfolio');
 
 $travels = $bdd->query('SELECT * FROM categories_travel INNER JOIN travels ON categories_travel.id = travels.category_id');
 
-$categories_travel = $bdd->query('SELECT * FROM categories_travel');
+$categories_travel_select = $bdd->query('SELECT * FROM categories_travel');
+
+$categories_travel_add = $bdd->query('SELECT * FROM categories_travel');