+++ admin/assets/manageTravels/content-travels.php
@@ -5,11 +5,11 @@
 <form action="#" method="GET">
 <select name="select_category">
         <option value="default"> ----- </option>
-        <?php while ($donnees = $categories_travel->fetch()) { ?>
+        <?php while ($donnees = $categories_travel_select->fetch()) { ?>
             <option value="<?php echo $donnees['category']; ?>"><?php echo $donnees['category'] ?></option>
         <?php } ?>
     </select>
-    <input type="submit" value="Submit">
+    <input type="submit" value="Select">
 </form>
 
 <section>
@@ -32,7 +32,9 @@
                 <tr>
                     <td><?php echo $donnees['category']; ?></td>
                     <td><?php echo $donnees['picture']; ?></td>
-                    <td><?php echo $donnees['pictures_id'] ?></td>
+                    <td>
+                        <?php echo $donnees['picture_1']; ?>
+                    </td>
                     <td><?php echo $donnees['title']; ?></td>
                     <td class="text-justify"><?php echo $donnees['description']; ?></td>
                     <td><?php echo $donnees['display_carousel']; ?></td>
@@ -49,18 +51,33 @@
             <div>
                 <select name="category">
                     <option value="0">---</option>
-                    <?php while ($donnees = $categories_travel->fetch()) { ?>
+                    <?php while ($donnees = $categories_travel_add->fetch()) { ?>
                         <option value="<?php echo $donnees['id']; ?>"><?php echo $donnees['category']; ?></option>
                     <?php } ?>
                 </select>
             </div>
             <div>
-                <label for="file" class="label-file-input">Choose preview picture</label>
+                <label for="file" class="label-file-input-preview">Choose preview picture</label>
                 <input id="file" type="file" name="picture" style="display: none;">
             </div>
-            <div>
-                <label for="file" class="label-file-input">Choose picture 1</label>
-                <input id="file" type="file" name="picture1" style="display: none;">
+            <div class="choose-pictures-input">
+                <a href="javascript:void(0)" class="add-pictures-link">Add pictures</a>
+                <div class="add-pictures-input">
+                    <hr>
+                    <div>
+                        <label for="file" class="label-file-input">Choose picture</label>
+                        <input id="file" type="file" name="picture_1" style="display: none;">
+                    </div>
+                    <div>
+                        <label for="file" class="label-file-input">Choose picture</label>
+                        <input id="file" type="file" name="picture_2" style="display: none;">
+                    </div>
+                    <div>
+                        <label for="file" class="label-file-input">Choose picture</label>
+                        <input id="file" type="file" name="picture_3" style="display: none;">
+                    </div>
+                    <hr>
+                </div>
             </div>
             <div>
                 <input type="text" name="title" placeholder="Title">